# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tools/{toolId}/secrets": {
      "post": {
        "description": "Securely store encrypted credentials for an external tool. These credentials will be used when the tool is executed in workflows.",
        "operationId": "ToolSecretsController_storeCredentials",
        "parameters": [
          {
            "name": "toolId",
            "required": true,
            "in": "path",
            "description": "Tool ID to store credentials for",
            "schema": {
              "example": "tool_slack_123",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Tool credentials to store securely",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "credentials": {
                    "type": "object",
                    "description": "Key-value pairs of credentials (will be encrypted)",
                    "example": {
                      "apiKey": "your-api-key-here",
                      "webhookUrl": "https://hooks.slack.com/services/YOUR/WEBHOOK/URL",
                      "botToken": "your-bot-token-here"
                    }
                  }
                },
                "required": [
                  "credentials"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Credentials stored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "toolId": {
                      "type": "string",
                      "example": "tool_slack_123"
                    },
                    "toolName": {
                      "type": "string",
                      "example": "Slack Integration"
                    },
                    "maskedCredentials": {
                      "type": "object",
                      "example": {
                        "apiKey": "your-***...***key",
                        "webhookUrl": "https://***...***XX",
                        "botToken": "your-***...***token"
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials data"
          },
          "404": {
            "description": "Tool not found"
          }
        },
        "security": [
          {
            "x-org-id": []
          }
        ],
        "summary": "Store Tool Credentials",
        "tags": [
          "Tool Secrets"
        ]
      },
      "get": {
        "description": "Retrieve stored credentials for a tool. By default returns masked values for security. Use unmask=true to get actual values (use carefully).",
        "operationId": "ToolSecretsController_getCredentials",
        "parameters": [
          {
            "name": "toolId",
            "required": true,
            "in": "path",
            "description": "Tool ID to retrieve credentials for",
            "schema": {
              "example": "tool_slack_123",
              "type": "string"
            }
          },
          {
            "name": "unmask",
            "required": false,
            "in": "query",
            "description": "Whether to return unmasked credential values (use with extreme caution)",
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Credentials retrieved successfully (masked by default)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "title": "Masked Credentials Response",
                      "type": "object",
                      "properties": {
                        "toolId": {
                          "type": "string",
                          "example": "tool_slack_123"
                        },
                        "toolName": {
                          "type": "string",
                          "example": "Slack Integration"
                        },
                        "maskedCredentials": {
                          "type": "object",
                          "example": {
                            "apiKey": "your-***...***key",
                            "webhookUrl": "https://***...***XX",
                            "botToken": "your-***...***token"
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    {
                      "title": "Unmasked Credentials Response (unmask=true)",
                      "type": "object",
                      "properties": {
                        "toolId": {
                          "type": "string",
                          "example": "tool_slack_123"
                        },
                        "toolName": {
                          "type": "string",
                          "example": "Slack Integration"
                        },
                        "credentials": {
                          "type": "object",
                          "example": {
                            "apiKey": "your-api-key-here",
                            "webhookUrl": "https://hooks.slack.com/services/YOUR/WEBHOOK/URL",
                            "botToken": "your-bot-token-here"
                          }
                        },
                        "maskedCredentials": {
                          "type": "object",
                          "example": {
                            "apiKey": "your-***...***key",
                            "webhookUrl": "https://***...***XX",
                            "botToken": "your-***...***token"
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Tool credentials not found"
          }
        },
        "security": [
          {
            "x-org-id": []
          }
        ],
        "summary": "Get Tool Credentials",
        "tags": [
          "Tool Secrets"
        ]
      },
      "delete": {
        "description": "Permanently delete stored credentials for a tool. This will affect any workflows that use this tool.",
        "operationId": "ToolSecretsController_deleteCredentials",
        "parameters": [
          {
            "name": "toolId",
            "required": true,
            "in": "path",
            "description": "Tool ID to delete credentials for",
            "schema": {
              "example": "tool_slack_123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Credentials deleted successfully"
          },
          "404": {
            "description": "Tool credentials not found"
          },
          "409": {
            "description": "Cannot delete credentials that are being used by active workflows"
          }
        },
        "security": [
          {
            "x-org-id": []
          }
        ],
        "summary": "Delete Tool Credentials",
        "tags": [
          "Tool Secrets"
        ]
      }
    },
    "/tools/secrets": {
      "get": {
        "description": "Get all tools in the organization with their credential status. Shows which tools have credentials stored and which ones need setup.",
        "operationId": "ToolSecretsListController_listToolsWithCredentials",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved tools with credential status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "toolId": {
                        "type": "string",
                        "example": "tool_slack_123"
                      },
                      "toolName": {
                        "type": "string",
                        "example": "Slack Integration"
                      },
                      "toolType": {
                        "type": "string",
                        "example": "notification"
                      },
                      "hasCredentials": {
                        "type": "boolean",
                        "example": true,
                        "description": "Whether credentials are stored for this tool"
                      },
                      "credentialKeys": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        },
                        "example": [
                          "apiKey",
                          "webhookUrl",
                          "botToken"
                        ],
                        "description": "List of credential keys (if credentials exist)"
                      },
                      "lastUpdated": {
                        "type": "string",
                        "format": "date-time",
                        "description": "When credentials were last updated (if they exist)"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "x-org-id": []
          }
        ],
        "summary": "List Tools with Credentials",
        "tags": [
          "Tool Secrets List"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tolstoy API",
    "description": "Interactive API reference for Tolstoy workflow automation platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://tolstoy.getpullse.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-org-id",
        "description": "Organization ID for multi-tenant access"
      },
      "x-user-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "User ID for request context"
      }
    },
...
  }
}